refactor(tests): move 'MockTime' to common
should be in a common 'testutils' module

but creating a module for a class seems overkill

allows 'JvmTest' to depend on 'MockTime'
