+++ AnkiDroid/src/test/java/com/ichi2/testutils/JvmTest.kt
@@ -19,6 +19,7 @@ package com.ichi2.testutils
 import android.annotation.SuppressLint
 import androidx.annotation.CallSuper
 import com.ichi2.anki.CollectionManager
+import com.ichi2.anki.common.time.MockTime
 import com.ichi2.anki.common.time.TimeManager
 import com.ichi2.anki.libanki.Collection
 import com.ichi2.anki.libanki.Storage
